5.3

+++ src/Illuminate/Cache/DatabaseStore.php
@@ -92,7 +92,7 @@ class DatabaseStore implements Store
      *
      * @param  string  $key
      * @param  mixed   $value
-     * @param  int     $minutes
+     * @param  float|int  $minutes
      * @return void
      */
     public function put($key, $value, $minutes)
@@ -104,7 +104,7 @@ class DatabaseStore implements Store
         // time and place that on the table so we will check it on our retrieval.
         $value = $this->encrypter->encrypt($value);
 
-        $expiration = $this->getTime() + ($minutes * 60);
+        $expiration = $this->getTime() + (int) ($minutes * 60);
 
         try {
             $this->table()->insert(compact('key', 'value', 'expiration'));
@@ -128,7 +128,7 @@ class DatabaseStore implements Store
     }
 
     /**
-     * Increment the value of an item in the cache.
+     * Decrement the value of an item in the cache.
      *
      * @param  string  $key
      * @param  mixed   $value
@@ -165,7 +165,7 @@ class DatabaseStore implements Store
             }
 
             $current = $this->encrypter->decrypt($cache->value);
-            $new = $callback($current, $value);
+            $new = $callback((int) $current, $value);
 
             if (! is_numeric($current)) {
                 return false;
